Fix/energy accumulation precision (#64)
* fix: change energy values from float to double to fix sum issues and for improved precision in large energy calculations

* fix: implementation of new double in all other places (modbus, api)

* fix: implement backward compatibility for energy values migration from float to double

* fix: increase log queue sizes for improved logging capacity

* fix: update default configuration gain values to calibrated settings

* fix: update roundToDecimals function to return double and adjust Modbus handling for energy values

+++ source/src/customserver.cpp
@@ -1835,11 +1835,11 @@ namespace CustomServer
                 }
 
                 uint8_t channel = doc["channel"].as<uint8_t>();
-                float activeEnergyImported = doc["activeEnergyImported"].as<float>();
-                float activeEnergyExported = doc["activeEnergyExported"].as<float>();
-                float reactiveEnergyImported = doc["reactiveEnergyImported"].as<float>();
-                float reactiveEnergyExported = doc["reactiveEnergyExported"].as<float>();
-                float apparentEnergy = doc["apparentEnergy"].as<float>();
+                double activeEnergyImported = doc["activeEnergyImported"].as<double>();
+                double activeEnergyExported = doc["activeEnergyExported"].as<double>();
+                double reactiveEnergyImported = doc["reactiveEnergyImported"].as<double>();
+                double reactiveEnergyExported = doc["reactiveEnergyExported"].as<double>();
+                double apparentEnergy = doc["apparentEnergy"].as<double>();
 
                 if (Ade7953::setEnergyValues(channel, activeEnergyImported, activeEnergyExported, 
                                            reactiveEnergyImported, reactiveEnergyExported, apparentEnergy))
